Boss room dialogue
Boss Rush room:
Added Tokaron's dialogue from the PSP version
Slightly adjusted Prince's dialogue to match PSP/PC JP
Added Giga's dialogue from PSP version, but you can't actually speak to him. Parin makes a joke about the collision being different in the PC version, so that seems to be the real reason (in the PSP version, you have to talk to him at an angle, so the collision probably is janky/not working on PC). Not sure if this can be fixed, since of course no one knows how to edit collision yet (probably something to do with the .exe)... Here's the dialogue.

> 	"Giiiiigaaaa."
> 	"That's strange...","You couldn't talk in the original PC version.","Maybe the collision detection is different..."
> 	"Giga giga..."
> 	"Poor thing.","You can only talk in one place,","and that's all you get to say..."
> 	""You called，master?""
> 	"That was your line，right?"
> 	"Giga giga..."

Changed Roger's name to his JP name, Lassi (ラッシー)

Changed Lassi's "We love Gurumin" dialogue (made up by Mastiff because the original PC dialogue was about the video card the devs used) to the PSP dialogue